Add definitions and types for incidents, users
This is a bit of a slog PR adding a number of types each with a Schema
and Serialize method.

We want to control what we add to the schema, and what we then write
out. Some fields are deprecated (not a lot) and this gives us control.

This will currently export all incident fields for the incidents v2 API
and their nested structures.