Add the discovery flag and method, along with model package
We want to be able to output schemas when we run in discovery mode so
this PR adds that flag along with the associated method to do so.

It just goes through the streams and calls Schema() along with adding a
few other fields. In the future we will also need to generate metadata
as part of that (and consume the file back in).

The other change here is the start of pulling out the models we use to
generate schemas and serialize data into their own package.

This PR just pulls out the schema itself - but a followup will add a
bunch of models representing entities in our API that we expose. Will
just follow a simple pattern of having a schema / serialize method on
them.