Add basic stream level metadata to catalog entries
This PR does a few big changes, mainly adding the concept of metadata to
the catalog and it's entries.

We create and emit a few default fields into entries now that let us
specify how we will act for a given stream - users can also edit these
values (along with adding others) so we need to read that catalog file
back in if specified.

When reading it in we parse it back into our catalog object and then use
that to try and filter the streams we will act on.

The options are a little weird here but basically, a user can set
"selected" (we cannot) if the selected is not set, then we can check if
"selected-by-default" is set, as this is something we typically control
and will have turned on.

If none of those match or are false then we don't include the stream.

If we do not find a catalog file at all - then we will create our
default one (same as running --discover) and just use that immediately.