Emit column level metadata in catalog and consume it back in
We want to allow users to configure which columns they ingest or don't
ingest in and out of their incident data.

This PR creates column level metadata for each property in a schema (it
doesn't do nested properties).

This lets us say yay or nay on a column by inspecting the metadata and
checking the typical "selected" or "selected-by-default" properties we
have there.