Add a stream filter for ... filtering streams
This implements the same interface as every stream and acts as a filter
on the stream it wraps allowing us to nicely filter out disabled fields
from both schemas and their results.

We could probably get more clever on the result generation to save
having to loop over again but for now this works yay!