Fix invalid schemas and timestamp formatting
While testing running this tap against stitch itself there were a few
validations that cropped up - mainly around timestamps that actually
need to be optional but are not.

There was a typo on another field also which has been corrected.

The final change was to the timestamp for when a record has been
emitted. The stitch documentation says:

```
This should be an RFC3339 formatted date-time, like "2017-11-20T16:45:33.000Z".
```

What they don't say is that their python library demands it's locked to UTC e.g.
you cannot have an actually valid RFC3339 timestamp such as:

`time_extracted=2023-10-22T19:28:41+01:00`

which has the timezone present. Why is this? Who knows - it's a
timestamp, it's valid, it has timezone information in it that would let
you calculate UTC.

So for now lets just emit with things cast to UTC first.