Add documentation for stitch contribution
This is documentation stitch require you to have around the tap that
will show on the website page about your app when users click for
detail.

It details how to get set up, the type of replication we do, and the
schemas we will create (including links to our API docs).