Add an integration test for discovery using the compiled binary
We want to be able to make sure we do not break anything in the tap
in terms of the schemas we output along with the records we expect to
output.

This is the first stage of this which compiles the tap in the test into
a binary that we can then run commands against. It just runs the
discovery command at the moment - but we can check the output against a
committed file and regenerate it if required.